+++ src/extensionApi.ts
@@ -25,6 +25,7 @@ export interface ResponseOutputProcessor {
 export interface DocumentStore {
   readonly httpFileStore: httpyac.store.HttpFileStore;
   activeEnvironment: Array<string> | undefined;
+  getActiveEnvironment(httpFile: httpyac.HttpFile): Array<string> | undefined;
   getDocumentPathLike: (document: vscode.TextDocument) => httpyac.PathLike;
   getHttpFile(document: vscode.TextDocument): Promise<httpyac.HttpFile | undefined>;
   getAll(): Array<httpyac.HttpFile>;