+++ src/utils/httpRegionUtils.ts
@@ -20,7 +20,7 @@ export async function getHttpRegionFromLine(
       );
       if (httpRegion) {
         return {
-          activeEnvironment: httpFile.activeEnvironment || documentStore.activeEnvironment,
+          activeEnvironment: documentStore.getActiveEnvironment(httpFile),
           httpRegion,
           httpFile,
         };